Add a projection function to MapSVG
The projection function determines, how geometries are projected onto a 2D plane. Inside the class, if a geometry is added to the canvas, this projection together with a linear transformation to the svg coordinate system computes the geometries coordinates. This makes it very consistent how geometries can be added to the canvas (always in lon/lat).