Fix string representation of MapSVG
The string representation of the parent class svg.SVG included every instance attribute into the SVG-string for the xml file. This included instance attributes in the MapSVG that I set (e.g. projection), which destroyed the SVG-file. This is now fixed (even though the fix is a little ugly, but it works fine).