Add clipping of geometries before orthographic projection
I had a lot of trouble with invalid shapes (with inf) after using the orthographic projection. The solution is now to clip geometries to visible areas BEFORE projection. This is still a little bit tricky, since we can only approximate the shape that is visible and it looks a little bit hacky sometimes. But for now, it works well and can be used to circumvent some problems.