Use new approach in _get_visible_lon_lat
In the old approach, I was tinkering around with the coordinate values. I tried to include new points to make the shape touch the boundaries, and removed points. This was very unclean and didn't work properly.

Now I work more with shapely geometries. I create geometries, merge them, shift them, and in the end create an approximation of the visible longitudes and latitudes. This works better (even though I don't know if its cleaner) and is maybe easier understandable and maintainable. But most importantly it works.