Exchange colored shadow by only black shadow
I did not like the look of the new shadow using the Lambertian for changing colors. I also didn't like using different shadows for land and sea, and they interfered with each other.

Therefore, this new shadow is only black with the opacity changing towards the edges. However, this is still different from the old shadow, since I still use the Lambertian computation to calculate the opacity values, instead of using a simple linear increase.