Change how river branches are snapped to their main
Before, the function went through start and end points and snapped them to a close point of another geometry if it was closer than a certain threshold.
In the new function, I use the information that is provided by the data, which has the destination of each river in the DataFrame. Therefore, now the closest point of the river is snapped to the closest point of the destination.